Split summary into markdown and markup fields.
summary now resembles the content field. The only realy difference being
that summary_markdown has a max_length of 765 characters. The limit is
not present on summary_markup as the html markup will be longer than
it's markdown counter part.

This limit may be lowered in the future if I find 765 characters to be
too long.