Switched to Django views for page switching
Previously was using Polymer's core-animated-pages to switch between the
home, projects, about, and contact pages. It works well but it seems to
be more webapp orientated. I've decided this blog is better served with
a more traditional style of navigation and haved moved towards that with
these changes. Poylmer elements will still be in use for certain things
because I like the aesthetic.